backend/jaeger/requirements.txt
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-exporter-jaeger==1.12.0
